--- dummy.json
@@ -1,86 +1,0 @@
-{
-   "paragraph":"Op het groene veld vol passie, waar de voetbalhelden strijden. Schijnbeweging, tact en actie, het publiek kan zich verblijden. Waar de bal de netten treft, stijgt het gejuich als een gedicht. Voetbal passie steeds gehoopt, voel de kracht in dit vuurlicht.",
-   "keywords":[
-      {
-         "keyword":"voetbalhelden",
-         "alternatives":[
-            "stervoetballers",
-            "voetbalkampioenen",
-            "topspelers",
-            "voetbaltalenten"
-         ]
-      },
-      {
-         "keyword":"schijnbeweging",
-         "alternatives":[
-            "finte",
-            "truc",
-            "dribbel",
-            "spelinzicht"
-         ]
-      },
-      {
-         "keyword":"tact",
-         "alternatives":[
-            "strategie",
-            "slimheid",
-            "inzicht",
-            "techniek"
-         ]
-      },
-      {
-         "keyword":"actie",
-         "alternatives":[
-            "beweging",
-            "dynamiek",
-            "snelheid",
-            "energie"
-         ]
-      },
-      {
-         "keyword":"publiek",
-         "alternatives":[
-            "fans",
-            "supporters",
-            "aanhang",
-            "toeschouwers"
-         ]
-      },
-      {
-         "keyword":"netten",
-         "alternatives":[
-            "doel",
-            "goal",
-            "score",
-            "succes"
-         ]
-      },
-      {
-         "keyword":"gejuich",
-         "alternatives":[
-            "gezangen",
-            "applaus",
-            "ovatie",
-            "enthousiasme"
-         ]
-      },
-      {
-         "keyword":"passie",
-         "alternatives":[
-            "bevlogenheid",
-            "hartstocht",
-            "enthousiasme",
-            "toewijding"
-         ]
-      },
-      {
-         "keyword":"vuurlicht",
-         "alternatives":[
-            "energie",
-            "gloed",
-            "warmte",
-            "verlichting"
-         ]
-      }
-   ]
-}
