+++ package.json
@@ -21,19 +21,11 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@fastify/auth": "^4.2.0",
-    "@fastify/bearer-auth": "^9.0.0",
-    "@fastify/cookie": "^8.3.0",
-    "@fastify/formbody": "^7.4.0",
-    "@fastify/rate-limit": "^8.0.0",
-    "@fastify/static": "^6.10.0",
-    "@fastify/view": "^7.4.1",
     "dotenv": "^16.0.3",
     "eta": "^2.0.1",
     "express": "^4.18.2",
     "express-handlebars": "^7.0.6",
     "express-rate-limit": "^6.7.0",
-    "fastify": "^4.15.0",
     "node-appwrite": "^8.2.0",
     "node-fetch": "^2.6.9",
     "openai": "^3.2.1"
