--- src/js/components/index.js
@@ -1,3 +1,0 @@
-// Importeer folder
-// index.js bestand word geimporteerd
-import "./test/";
