--- src/js/components/test/index.js
@@ -1,1 +1,0 @@
-import "./test.js";
