--- src/js/components/test/test.js
@@ -1,1 +1,0 @@
-console.log("test.js is geladen");
