+++ src/js/home.js
@@ -1,0 +1,6 @@
+const home = document.querySelector("[data-home]");
+const homeBtn = document.querySelector("[data-home-btn]");
+
+homeBtn.addEventListener("click", () => {
+  home.classList.toggle("home--visible");
+});
