+++ src/js/themeBtn.js
@@ -1,0 +1,9 @@
+theme.addEventListener("input", (e) => {
+  if (e.target.value !== "") {
+    submit.classList.add("cta--visible");
+    submit.classList.remove("cta--hidden");
+  } else {
+    submit.classList.remove("cta--visible");
+    submit.classList.add("cta--hidden");
+  }
+});
