+++ src/scss/components/_all.scss
@@ -1,4 +1,3 @@
-// Import het 'all' bestand uit de component map
 @import "control/all";
 @import "cta/all";
 @import "icon/all";
@@ -11,5 +10,4 @@
 @import "page/all";
 @import "text/all";
 @import "title/all";
-@import "test/all";
 @import "ornaments/all";
