+++ src/scss/components/cta/_cta.scss
@@ -40,6 +40,7 @@
     margin: 2em 0;
     text-decoration: none;
     color: $tet-color;
+    background-color: transparent;
 
     span {
       color: inherit;
