+++ src/scss/components/icon/_all.scss
@@ -1,0 +1,1 @@
+@import "icon";
