+++ src/scss/components/ornaments/_ornaments.scss
@@ -1,3 +1,5 @@
+$offset: 10px;
+
 .ornaments {
   span {
     position: absolute;
@@ -7,25 +9,25 @@
     background-size: contain;
 
     &:first-of-type {
-      top: 0;
-      left: 0;
+      top: $offset;
+      left: $offset;
       transform: rotate(-90deg);
     }
 
     &:nth-of-type(2) {
-      top: 0;
-      right: 0;
+      top: $offset;
+      right: $offset;
     }
 
     &:nth-of-type(3) {
-      bottom: 0;
-      right: 0;
+      bottom: $offset;
+      right: $offset;
       transform: rotate(90deg);
     }
 
     &:nth-of-type(4) {
-      bottom: 0;
-      left: 0;
+      bottom: $offset;
+      left: $offset;
       transform: rotate(-180deg);
     }
   }
