+++ src/scss/components/page/_page.scss
@@ -11,7 +11,8 @@
   height: 100%;
   transform: translateY(-100%);
   transition: all 300ms ease-in-out;
-
+  background-color: $bg-color;
+  border: 10px solid black;
   &--visible {
     transform: translateY(0);
   }
