--- src/scss/components/test/_all.scss
@@ -1,2 +1,0 @@
-// Import all components in this folder
-@import "test";
