--- src/scss/components/test/_test.scss
@@ -1,6 +1,0 @@
-// Styling voor de .test (block) component
-.test {
-  color: red;
-  background-color: green;
-  border: 1px solid orange;
-}
