+++ src/scss/settings/_color.scss
@@ -4,4 +4,6 @@ $sec-color: orange;
 $tet-color: black;
 $qua-color: antiquewhite;
 
+$bg-color: #faebd7;
+
 // Scss variablen
