--- views/details.eta
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html data-theme="light" lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <script src="https://unpkg.com/htmx.org@1.8.6"></script>
-
-    <link rel="stylesheet" href="public/css/all.css" />
-    <script src="public/scripts/index.js" type="module"></script>
-
-    <title><%= it.title %></title>
-  </head>
-  <body>
-    <%~ includeFile('partials/header') %>
-      <main class="container page">
-      	<h1 class="title title--h1">Result</h1>
-        <p class="text text--big">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut id massa maximus, consequat lorem
-          et, malesuada metus. Donec facilisis porttitor nisi, gravida ornare libero. Donec cursus leo vel risus
-          tincidunt, eu commodo diam elementum. Integer ullamcorper elit orci, quis mollis erat venenatis dictum.
-          Mauris orci quam, fermentum eget orci sed, semper consequat est. Proin a luctus est. Praesent faucibus ante
-          finibus, cursus ante at, vestibulum sapien. Praesent vitae mi vel lorem ultricies iaculis. Sed vitae
-          fringilla tellus. Cras cursus enim nec lobortis cursus. Etiam bibendum ante in est mollis, at pretium lectus
-          placerat. Curabitur felis purus, tincidunt eu euismod eget, efficitur vitae metus. Duis fringilla elit et
-          fermentum tempus. Morbi sollicitudin libero nibh, auctor pulvinar augue convallis a.
-        </p>
-		    <button class="cta" onClick="showOptions()"">Options</button>
-    </main>
-  </body>
-</html>
