--- views/error.eta
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html data-theme="light" lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="https://unpkg.com/@picocss/pico@1.*/css/pico.min.css">
-    <script src="https://unpkg.com/htmx.org@1.8.6"></script>
-    <title><%= it.title %></title>
-</head>
-<body>
-<main class="container">
-    <nav aria-label="breadcrumb">
-        <ul>
-            <li><a href="/">Home</a></li>
-            <li><a href="/docs">Docs</a></li>
-        </ul>
-    </nav>
-
-    <hgroup>
-        <h1>404</h1>
-        <h2>Page not found</h2>
-    </hgroup>
-</main>
-</body>
-</html>
