+++ views/home.hbs
@@ -1,5 +1,4 @@
-<dialog class="home" data-home>
-  {{! add home--visible to show the home screen }}
+<dialog class="home home--visible" data-home>
   <h1 class="title title--h0">De Stads Dichter</h1>
 
   <button class="cta cta__start" data-home-btn>
