--- views/index.eta
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html data-theme="light" lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <script src="https://unpkg.com/htmx.org@1.8.6"></script>
-
-    <link rel="stylesheet" href="public/css/all.css" />
-    <script src="public/scripts/index.js" type="module"></script>
-
-    <title>Home | <%= it.title %></title>
-  </head>
-  <body>
-    <%~ includeFile('partials/header') %>
-  </body>
-</html>
