--- views/output.eta
@@ -1,35 +1,0 @@
-<% const paragraph = it.data.paragraph; %>
-<% const keywords = it.data.keywords; %>
-<% const lines = paragraph.split("."); %>
-<% const wordsByLine = lines.map(line => line.split(" ")); %>
-<% const serial = wordsByLine.map(words => ({ words })); %>
-
-<% for (let i = 0; i < wordsByLine.length; i++) { %>
-<% for (let j = 0; j < wordsByLine[i].length; j++) { %>
-<% let word = wordsByLine[i][j]; %>
-<% let isBolded = false; %>
-<% for (let k = 0; k < it.data.keywords.length; k++) { %>
-<% let keyword = it.data.keywords[k].keyword; %>
-<% if (word === keyword && !isBolded) { %>
-
-<select hx-indicator="#spinner-2" name="<%= word %>" hx-get="/cont" hx-target="#output" style="display: inline-block; width: fit-content;">
-    <option value="<%= word %>" selected><%= word %></option>
-    <% for (let l = 0; l < it.data.keywords[k].alternatives.length; l++) { %>
-    <option value="<%= it.data.keywords[k].alternatives[l] %>"><%= it.data.keywords[k].alternatives[l] %></option>
-    <% } %>
-</select>
-
-<% isBolded = true; %>
-<% } %>
-<% } %>
-<% if (!isBolded) { %>
-<%= word + " "%>
-<% } %>
-<% } %>
-.<br>
-<% } %>
-
-
-
-
-
